requirements.txt
opencv-python==4.9.0.80
pytesseract==0.3.10
openai-whisper==20231117
Pillow==10.2.0
numpy==1.26.4
PyYAML==6.0.1
